requirements.txt
joblib==1.2.0
matplotlib==3.5.3
mlflow==2.15.1
mlflow_skinny==2.15.1
numpy==1.24.4
pandas==1.5.3
Pillow==10.4.0
pycaret==3.2.0
scikit_learn==1.2.2
shap==0.44.1
streamlit==1.31.1
streamlit_pandas_profiling==0.1.3
xgboost==1.7.6
ydata_profiling==4.9.0